requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.1
PyPDF2>=3.0.1
pytube>=15.0.0
python-dotenv>=1.0.0
six>=1.16.0
python-dateutil>=2.8.2
pandas>=1.3.0
pydantic>=2.5.2
pillow>=10.1.0
streamlit-extras>=0.3.5
streamlit-card>=0.0.61
youtube-transcript-api>=0.6.3
